typo #466
