+++ mldsa/field_arm64_test.go
@@ -6,68 +6,48 @@
 
 package mldsa
 
-import "testing"
+import (
+	mathrand "math/rand/v2"
+	"testing"
+)
 
-var benchmarkNTTMulArm64Sink nttElement
-
-func BenchmarkNTTMulArm64(b *testing.B) {
-	left := ntt(randomRingElement())
-	right := ntt(randomRingElement())
-	var out nttElement
-
-	b.ReportAllocs()
-
-	b.Run("into/generic", func(b *testing.B) {
-		for i := 0; i < b.N; i++ {
-			nttMulGeneric(&out, &left, &right)
+func TestNTTMulNEONMatchesGeneric(t *testing.T) {
+	for range 64 {
+		var lhs, rhs, got, want nttElement
+		for i := range lhs {
+			lhs[i] = fieldElement(mathrand.IntN(q))
+			rhs[i] = fieldElement(mathrand.IntN(q))
 		}
-		benchmarkNTTMulArm64Sink = out
-	})
 
-	b.Run("into/dispatch", func(b *testing.B) {
-		for i := 0; i < b.N; i++ {
-			nttMul(&out, &left, &right)
-		}
-		benchmarkNTTMulArm64Sink = out
-	})
+		nttMulNEON(&lhs, &rhs, &got)
+		nttMulGeneric(&want, &lhs, &rhs)
 
-	b.Run("into/neon", func(b *testing.B) {
-		for i := 0; i < b.N; i++ {
-			nttMulNEON(&out, &left, &right)
+		for i := range got {
+			if got[i] != want[i] {
+				t.Fatalf("index %d: got %d, want %d", i, got[i], want[i])
+			}
 		}
-		benchmarkNTTMulArm64Sink = out
-	})
+	}
 }
 
-func BenchmarkNTTMulAccArm64(b *testing.B) {
-	left := ntt(randomRingElement())
-	right := ntt(randomRingElement())
-	base := ntt(randomRingElement())
-	var acc nttElement
-
-	b.ReportAllocs()
-
-	b.Run("into/generic", func(b *testing.B) {
-		for i := 0; i < b.N; i++ {
-			acc = base
-			nttMulAccGeneric(&acc, &left, &right)
+func TestNTTMulAccNEONMatchesGeneric(t *testing.T) {
+	for range 64 {
+		var lhs, rhs, got, want nttElement
+		for i := range lhs {
+			lhs[i] = fieldElement(mathrand.IntN(q))
+			rhs[i] = fieldElement(mathrand.IntN(q))
+			// pre-populate accumulator with random values
+			got[i] = fieldElement(mathrand.IntN(q))
+			want[i] = got[i]
 		}
-		benchmarkNTTMulArm64Sink = acc
-	})
 
-	b.Run("into/dispatch", func(b *testing.B) {
-		for i := 0; i < b.N; i++ {
-			acc = base
-			nttMulAcc(&acc, &left, &right)
-		}
-		benchmarkNTTMulArm64Sink = acc
-	})
+		nttMulAccNEON(&lhs, &rhs, &got)
+		nttMulAccGeneric(&want, &lhs, &rhs)
 
-	b.Run("into/neon", func(b *testing.B) {
-		for i := 0; i < b.N; i++ {
-			acc = base
-			nttMulAccNEON(&left, &right, &acc)
+		for i := range got {
+			if got[i] != want[i] {
+				t.Fatalf("index %d: got %d, want %d", i, got[i], want[i])
+			}
 		}
-		benchmarkNTTMulArm64Sink = acc
-	})
+	}
 }
